<glide> Rollforward [] with a bit to track scale vs non-scale transformations.
This allows us to apply scale transformations if we decode from a video but still use the hardware config if we decode from an image. Unfortunately it won't work for custom transformations that only scale. We'd probably have to extend the Transformation API to let transformations indicate that they only scale and I'm not sure that that's common enough to be useful.